refactor: move http basic response methods to chatui
