+++ packages/packages.json
@@ -16,15 +16,15 @@
         "contract/valory/relayer/0.1.0": "bafybeihypuljybkocl6iiacy52py7i5iqxdli3ily66q7b3nego4qajvne",
         "contract/valory/market_maker/0.1.0": "bafybeigzfyzbixum5cqvcbyc24dqwce25dr6nyqvgjm2eqpt6tgo2perd4",
         "skill/valory/market_manager_abci/0.1.0": "bafybeigzldcghoxbg73e53ha2v3cayshswrrgeg4bqpcwd6al237v2ssle",
-        "skill/valory/decision_maker_abci/0.1.0": "bafybeifrrwylo6fedhfuz7p3bo3rrvr5trt6oxhlkzfylxpnlcwmgywlie",
-        "skill/valory/trader_abci/0.1.0": "bafybeiaex3hdxnpwj2tilguzdug7qh2itmsxqnmob7z7uo4flymre4qgme",
-        "skill/valory/tx_settlement_multiplexer_abci/0.1.0": "bafybeidgjznmwcam6bnlhh7q474f2hzq4636fuxrbiu3kts22l37qu3v6m",
+        "skill/valory/decision_maker_abci/0.1.0": "bafybeibucsfg3bstu6epz7ygsuff6pdr4edurjvhkqjmqug3qpzyzgkw4y",
+        "skill/valory/trader_abci/0.1.0": "bafybeibrxe4eeqr5n4gc2sahsonljzcdiy7qwq6irexpza546hq74cjbua",
+        "skill/valory/tx_settlement_multiplexer_abci/0.1.0": "bafybeib4gjbqofobdzr5ukdkjkqna574c5fzabp3gscjdspobly3d6poqq",
         "skill/valory/staking_abci/0.1.0": "bafybeifupwkfbxa4c4jogpudvzwt5rkgfuedgd65sj2bf2z5ver4phq64m",
         "skill/valory/check_stop_trading_abci/0.1.0": "bafybeignl5w4o4ghcpan7nsor5elvuid4si7t4rnfqga47bwvlgp22xwsa",
-        "skill/valory/chatui_abci/0.1.0": "bafybeielfqzdrsbogofja7vmttxnahsao33ihhdx32jnw47xzc5uuzm3im",
-        "agent/valory/trader/0.1.0": "bafybeig34tv6hqipiwp44oajwuh6dx3lhzvlzcqzoqvztttjxb4rjr4pza",
-        "service/valory/trader/0.1.0": "bafybeihxp3clzn4j6lg5wju2pzzwx4n4zr5zni6mvv4twrl7sqbq7fx4xe",
-        "service/valory/trader_pearl/0.1.0": "bafybeichdog3btenchrzl2zs7nwdfi5zzjywc2j356du6ba6dzsinzyfwe"
+        "skill/valory/chatui_abci/0.1.0": "bafybeigstoa636yeqekwbg5ggvcjsuwhpec2u5epqvai6k33cwnvj3yqoa",
+        "agent/valory/trader/0.1.0": "bafybeigubokj24jutdvuf36ial5riktwdmmpba345tjt6amvtifv4l2da4",
+        "service/valory/trader/0.1.0": "bafybeiajgkw4yzc6h773u6yl6nklbzcsiabt5yzpayb7sqlkgxh4kmmoj4",
+        "service/valory/trader_pearl/0.1.0": "bafybeiao5xvisg6ocjhby6y4nvjsyac3iltesg6io43rtetqlrokzcx4ai"
     },
     "third_party": {
         "protocol/valory/acn_data_share/0.1.0": "bafybeih5ydonnvrwvy2ygfqgfabkr47s4yw3uqxztmwyfprulwfsoe7ipq",
