+++ packages/valory/services/trader_pearl/service.yaml
@@ -8,7 +8,7 @@ license: Apache-2.0
 fingerprint:
   README.md: bafybeibg7bdqpioh4lmvknw3ygnllfku32oca4eq5pqtvdrdsgw6buko7e
 fingerprint_ignore_patterns: []
-agent: valory/trader:0.1.0:bafybeig34tv6hqipiwp44oajwuh6dx3lhzvlzcqzoqvztttjxb4rjr4pza
+agent: valory/trader:0.1.0:bafybeigubokj24jutdvuf36ial5riktwdmmpba345tjt6amvtifv4l2da4
 number_of_agents: 1
 deployment:
   agent:
