+++ packages/valory/skills/chatui_abci/handlers.py
@@ -21,7 +21,8 @@
 
 import json
 from enum import Enum
-from typing import Any, Callable, Dict, List, Optional, cast
+from http import HTTPStatus
+from typing import Any, Callable, Dict, List, Optional, Union, cast
 from urllib.parse import urlparse
 
 from aea.configurations.data_types import PublicId
@@ -155,6 +156,137 @@ class HttpHandler(BaseHttpHandler):
                 (chatui_prompt_url, self._handle_chatui_prompt),
             ],
         }
+
+    def _send_http_response(
+        self,
+        http_msg: HttpMessage,
+        http_dialogue: HttpDialogue,
+        data: Union[str, Dict, List, bytes],
+        status_code: int,
+        status_text: str,
+        content_type: Optional[str] = None,
+    ) -> None:
+        """Generic method to send HTTP responses."""
+        headers = content_type or (
+            CONTENT_TYPES[".json"] if isinstance(data, (dict, list)) else DEFAULT_HEADER
+        )
+        headers += http_msg.headers
+
+        # Convert dictionary or list to JSON string
+        if isinstance(data, (dict, list)):
+            data = json.dumps(data)
+
+        http_response = http_dialogue.reply(
+            performative=HttpMessage.Performative.RESPONSE,
+            target_message=http_msg,
+            version=http_msg.version,
+            status_code=status_code,
+            status_text=status_text,
+            headers=headers,
+            body=data.encode("utf-8") if isinstance(data, str) else data,
+        )
+
+        self.context.logger.info("Responding with: {}".format(http_response))
+        self.context.outbox.put_message(message=http_response)
+
+    def _send_too_early_request_response(
+        self,
+        http_msg: HttpMessage,
+        http_dialogue: HttpDialogue,
+        data: Union[str, Dict, List, bytes],
+        content_type: Optional[str] = None,
+    ) -> None:
+        """Handle a HTTP too early request response."""
+        self._send_http_response(
+            http_msg,
+            http_dialogue,
+            data,
+            HTTPStatus.TOO_EARLY.value,
+            HTTPStatus.TOO_EARLY.phrase,
+            content_type,
+        )
+
+    def _send_too_many_requests_response(
+        self,
+        http_msg: HttpMessage,
+        http_dialogue: HttpDialogue,
+        data: Union[str, Dict, List, bytes],
+        content_type: Optional[str] = None,
+    ) -> None:
+        """Handle a HTTP too many requests response."""
+        self._send_http_response(
+            http_msg,
+            http_dialogue,
+            data,
+            HTTPStatus.TOO_MANY_REQUESTS.value,
+            HTTPStatus.TOO_MANY_REQUESTS.phrase,
+            content_type,
+        )
+
+    def _send_internal_server_error_response(
+        self,
+        http_msg: HttpMessage,
+        http_dialogue: HttpDialogue,
+        data: Union[str, Dict, List, bytes],
+        content_type: Optional[str] = None,
+    ) -> None:
+        """Handle a Http internal server error response."""
+        headers = content_type or (
+            CONTENT_TYPES[".json"] if isinstance(data, (dict, list)) else DEFAULT_HEADER
+        )
+        headers += http_msg.headers
+
+        # Convert dictionary or list to JSON string
+        if isinstance(data, (dict, list)):
+            data = json.dumps(data)
+
+        http_response = http_dialogue.reply(
+            performative=HttpMessage.Performative.RESPONSE,
+            target_message=http_msg,
+            version=http_msg.version,
+            status_code=HTTPStatus.INTERNAL_SERVER_ERROR.value,
+            status_text=HTTPStatus.INTERNAL_SERVER_ERROR.phrase,
+            headers=headers,
+            body=data.encode("utf-8") if isinstance(data, str) else data,
+        )
+
+        # Send response
+        self.context.logger.info("Responding with: {}".format(http_response))
+        self.context.outbox.put_message(message=http_response)
+
+    def _send_bad_request_response(
+        self,
+        http_msg: HttpMessage,
+        http_dialogue: HttpDialogue,
+        data: Union[str, Dict, List, bytes],
+        content_type: Optional[str] = None,
+    ) -> None:
+        """Handle a HTTP bad request."""
+        self._send_http_response(
+            http_msg,
+            http_dialogue,
+            data,
+            HTTPStatus.BAD_REQUEST.value,
+            HTTPStatus.BAD_REQUEST.phrase,
+            content_type,
+        )
+
+    def _send_ok_request_response(
+        self,
+        http_msg: HttpMessage,
+        http_dialogue: HttpDialogue,
+        data: Union[str, Dict, List, bytes],
+        content_type: Optional[str] = None,
+    ) -> None:
+        """Send an OK response with the provided data."""
+        self._send_http_response(
+            http_msg,
+            http_dialogue,
+            data,
+            HTTPStatus.OK.value,
+            "Success",
+            content_type,
+        )
 
     def _send_message(
         self,
@@ -297,7 +429,7 @@ class HttpHandler(BaseHttpHandler):
             self.context.logger.warning(
                 "No agent configuration update provided by the LLM."
             )
-            self._send_ok_response(
+            self._send_ok_request_response(
                 http_msg,
                 http_dialogue,
                 {
@@ -311,7 +443,7 @@ class HttpHandler(BaseHttpHandler):
             updated_agent_config
         )
 
-        self._send_ok_response(
+        self._send_ok_request_response(
             http_msg,
             http_dialogue,
             {
