+++ packages/valory/skills/decision_maker_abci/skill.yaml
@@ -111,7 +111,7 @@ skills:
 - valory/transaction_settlement_abci:0.1.0:bafybeic2ywzpwkyeqbzsvkbvurhsptemam4xtceihax2tmxmlxtgd3xpya
 - valory/mech_interact_abci:0.1.0:bafybeigkizx3ecohtyshr6tq55zvp4sl5cou344a2t6nvwfgund4xnrj4a
 - valory/staking_abci:0.1.0:bafybeifupwkfbxa4c4jogpudvzwt5rkgfuedgd65sj2bf2z5ver4phq64m
-- valory/chatui_abci:0.1.0:bafybeielfqzdrsbogofja7vmttxnahsao33ihhdx32jnw47xzc5uuzm3im
+- valory/chatui_abci:0.1.0:bafybeigstoa636yeqekwbg5ggvcjsuwhpec2u5epqvai6k33cwnvj3yqoa
 behaviours:
   main:
     args: {}
