+++ packages/valory/skills/trader_abci/handlers.py
@@ -21,9 +21,8 @@
 """This module contains the handlers for the 'trader_abci' skill."""
 
 import json
-from http import HTTPStatus
 from pathlib import Path
-from typing import Any, Dict, List, Optional, Union, cast
+from typing import Any, Dict, List, Optional, cast
 from urllib.parse import urlparse
 
 from aea.configurations.data_types import PublicId
@@ -150,137 +149,6 @@ class HttpHandler(BaseHttpHandler):
         """Get the appropriate content type header based on file extension."""
         return CONTENT_TYPES.get(file_path.suffix.lower(), DEFAULT_HEADER)
 
-    def _send_http_response(
-        self,
-        http_msg: HttpMessage,
-        http_dialogue: HttpDialogue,
-        data: Union[str, Dict, List, bytes],
-        status_code: int,
-        status_text: str,
-        content_type: Optional[str] = None,
-    ) -> None:
-        """Generic method to send HTTP responses."""
-        headers = content_type or (
-            CONTENT_TYPES[".json"] if isinstance(data, (dict, list)) else DEFAULT_HEADER
-        )
-        headers += http_msg.headers
-
-        # Convert dictionary or list to JSON string
-        if isinstance(data, (dict, list)):
-            data = json.dumps(data)
-
-        http_response = http_dialogue.reply(
-            performative=HttpMessage.Performative.RESPONSE,
-            target_message=http_msg,
-            version=http_msg.version,
-            status_code=status_code,
-            status_text=status_text,
-            headers=headers,
-            body=data.encode("utf-8") if isinstance(data, str) else data,
-        )
-
-        self.context.logger.info("Responding with: {}".format(http_response))
-        self.context.outbox.put_message(message=http_response)
-
-    def _send_too_early_request_response(
-        self,
-        http_msg: HttpMessage,
-        http_dialogue: HttpDialogue,
-        data: Union[str, Dict, List, bytes],
-        content_type: Optional[str] = None,
-    ) -> None:
-        """Handle a HTTP too early request response."""
-        self._send_http_response(
-            http_msg,
-            http_dialogue,
-            data,
-            HTTPStatus.TOO_EARLY.value,
-            HTTPStatus.TOO_EARLY.phrase,
-            content_type,
-        )
-
-    def _send_too_many_requests_response(
-        self,
-        http_msg: HttpMessage,
-        http_dialogue: HttpDialogue,
-        data: Union[str, Dict, List, bytes],
-        content_type: Optional[str] = None,
-    ) -> None:
-        """Handle a HTTP too many requests response."""
-        self._send_http_response(
-            http_msg,
-            http_dialogue,
-            data,
-            HTTPStatus.TOO_MANY_REQUESTS.value,
-            HTTPStatus.TOO_MANY_REQUESTS.phrase,
-            content_type,
-        )
-
-    def _send_internal_server_error_response(
-        self,
-        http_msg: HttpMessage,
-        http_dialogue: HttpDialogue,
-        data: Union[str, Dict, List, bytes],
-        content_type: Optional[str] = None,
-    ) -> None:
-        """Handle a Http internal server error response."""
-        headers = content_type or (
-            CONTENT_TYPES[".json"] if isinstance(data, (dict, list)) else DEFAULT_HEADER
-        )
-        headers += http_msg.headers
-
-        # Convert dictionary or list to JSON string
-        if isinstance(data, (dict, list)):
-            data = json.dumps(data)
-
-        http_response = http_dialogue.reply(
-            performative=HttpMessage.Performative.RESPONSE,
-            target_message=http_msg,
-            version=http_msg.version,
-            status_code=HTTPStatus.INTERNAL_SERVER_ERROR.value,
-            status_text=HTTPStatus.INTERNAL_SERVER_ERROR.phrase,
-            headers=headers,
-            body=data.encode("utf-8") if isinstance(data, str) else data,
-        )
-
-        # Send response
-        self.context.logger.info("Responding with: {}".format(http_response))
-        self.context.outbox.put_message(message=http_response)
-
-    def _send_bad_request_response(
-        self,
-        http_msg: HttpMessage,
-        http_dialogue: HttpDialogue,
-        data: Union[str, Dict, List, bytes],
-        content_type: Optional[str] = None,
-    ) -> None:
-        """Handle a HTTP bad request."""
-        self._send_http_response(
-            http_msg,
-            http_dialogue,
-            data,
-            HTTPStatus.BAD_REQUEST.value,
-            HTTPStatus.BAD_REQUEST.phrase,
-            content_type,
-        )
-
-    def _send_ok_response(
-        self,
-        http_msg: HttpMessage,
-        http_dialogue: HttpDialogue,
-        data: Union[str, Dict, List, bytes],
-        content_type: Optional[str] = None,
-    ) -> None:
-        """Send an OK response with the provided data."""
-        self._send_http_response(
-            http_msg,
-            http_dialogue,
-            data,
-            HTTPStatus.OK.value,
-            "Success",
-            content_type,
-        )
-
     def _handle_get_agent_info(
         self, http_msg: HttpMessage, http_dialogue: HttpDialogue
     ) -> None:
@@ -292,7 +160,7 @@ class HttpHandler(BaseHttpHandler):
             "service_id": self.staking_synchronized_data.service_id,
         }
         self.context.logger.info(f"Sending agent info: {data=}")
-        self._send_ok_response(http_msg, http_dialogue, data)
+        self._send_ok_request_response(http_msg, http_dialogue, data)
 
     def _handle_get_static_file(
         self, http_msg: HttpMessage, http_dialogue: HttpDialogue
@@ -323,7 +191,7 @@ class HttpHandler(BaseHttpHandler):
                 content_type = self._get_content_type(file_path)
 
                 # Send the file content as a response
-                self._send_ok_response(
+                self._send_ok_request_response(
                     http_msg, http_dialogue, file_content, content_type
                 )
             else:
@@ -336,7 +204,7 @@ class HttpHandler(BaseHttpHandler):
                     index_html = file.read()
 
                 # Send the HTML response
-                self._send_ok_response(http_msg, http_dialogue, index_html)
+                self._send_ok_request_response(http_msg, http_dialogue, index_html)
         except FileNotFoundError:
             self._handle_not_found(http_msg, http_dialogue)
