+++ packages/valory/skills/trader_abci/skill.yaml
@@ -12,7 +12,7 @@ fingerprint:
   composition.py: bafybeigl47wrvg6tpmslslqiwx63ugqhfqydjapycmfs35nvx77bqmc2ny
   dialogues.py: bafybeiaiv562rcf7cjenuowrus744jc2njycrobkl72jwebocuh3htx2mu
   fsm_specification.yaml: bafybeigoi5xhvgeekg7d37icw5xljqmay6fil453xavhhbszwamubqmeja
-  handlers.py: bafybeichwsqvskq56ozsmhgqohr5loj3kdnaalldf77fnmmqth352nsqqa
+  handlers.py: bafybeifux4qnekhrplmm7gqeyqlprziwli6fugd3eglmmc7y64qcddwvuy
   models.py: bafybeianrjqqbpci45d6dcoqcqs7gfyxqu26toncc2qv6zilpsz75hcd3e
   predict-ui-build/assets/agent-agentsfun-logo-Cl82UKSm.png: bafybeia3nla5pruiurpwvzex6bbsgset4k2foerma2iwy5zxmtjidur454
   predict-ui-build/assets/index-BrVYRq-p.js: bafybeid6cgvwnfytp3hf5c3kpykx5rewlbteopy72cvlt6vw2rkpkesrwm
@@ -38,8 +38,8 @@ skills:
 - valory/transaction_settlement_abci:0.1.0:bafybeic2ywzpwkyeqbzsvkbvurhsptemam4xtceihax2tmxmlxtgd3xpya
 - valory/termination_abci:0.1.0:bafybeibtbboau3q5fxfviwm7lbeix4z55uptfqqiyiu6siivxwkp3o5pju
 - valory/market_manager_abci:0.1.0:bafybeigzldcghoxbg73e53ha2v3cayshswrrgeg4bqpcwd6al237v2ssle
-- valory/decision_maker_abci:0.1.0:bafybeifrrwylo6fedhfuz7p3bo3rrvr5trt6oxhlkzfylxpnlcwmgywlie
-- valory/tx_settlement_multiplexer_abci:0.1.0:bafybeidgjznmwcam6bnlhh7q474f2hzq4636fuxrbiu3kts22l37qu3v6m
+- valory/decision_maker_abci:0.1.0:bafybeibucsfg3bstu6epz7ygsuff6pdr4edurjvhkqjmqug3qpzyzgkw4y
+- valory/tx_settlement_multiplexer_abci:0.1.0:bafybeib4gjbqofobdzr5ukdkjkqna574c5fzabp3gscjdspobly3d6poqq
 - valory/staking_abci:0.1.0:bafybeifupwkfbxa4c4jogpudvzwt5rkgfuedgd65sj2bf2z5ver4phq64m
 - valory/check_stop_trading_abci:0.1.0:bafybeignl5w4o4ghcpan7nsor5elvuid4si7t4rnfqga47bwvlgp22xwsa
 - valory/mech_interact_abci:0.1.0:bafybeigkizx3ecohtyshr6tq55zvp4sl5cou344a2t6nvwfgund4xnrj4a
